Delete element DONE
- Delete window added to admin window
- Fix bugs
- Optimisation
