+++ MyNetia/ConfirmationWindow.xaml
@@ -7,6 +7,7 @@
         mc:Ignorable="d"
         Style="{StaticResource window}"
         Title="MyNetia-Confirmation"
+        WindowStartupLocation="CenterOwner"
         MinHeight="250" MaxHeight="250"
         MinWidth="400" MaxWidth="400">
     <Window.Resources>
