--- MyNetia/RemoveWindow.xaml
@@ -1,107 +1,0 @@
-﻿<Window x:Class="MyNetia.RemoveWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:MyNetia"
-        mc:Ignorable="d"
-        Title="MyNetia-Remove"
-        Style="{StaticResource window}"
-        MinHeight="200" MaxHeight="200"
-        MinWidth="300" MaxWidth="300">
-    <Window.Resources>
-        <Style x:Key="labelTitleBar" BasedOn="{StaticResource labelBase}" TargetType="Label">
-            <Setter Property="FontSize" Value="20"/>
-            <Setter Property="Height" Value="30"/>
-            <Setter Property="Margin" Value="5, 0, 0, 0"/>
-        </Style>
-        
-        <Style x:Key="txtBox" BasedOn="{StaticResource txtBoxBase}" TargetType="TextBox">
-            <Setter Property="FontSize" Value="20"/>
-            <Setter Property="BorderThickness" Value="0"/>
-            <Setter Property="Height" Value="30"/>
-            <Setter Property="Width" Value="Auto"/>
-            <Setter Property="Margin" Value="5, 3, 5, 3"/>
-        </Style>
-        
-        <Style x:Key="txtBlock" BasedOn="{StaticResource txtBlockBase}" TargetType="TextBlock">
-            <Setter Property="FontSize" Value="18"/>
-            <Setter Property="Padding" Value="5"/>
-        </Style>
-    </Window.Resources>
-    <DockPanel>
-        <!--#region TITLE BAR-->
-        <Grid
-            x:Name="titleBar"
-            DockPanel.Dock="Top"
-            Style="{StaticResource gridTitleBar}"
-            MouseDown="titleBar_MouseDown">
-            <Grid.Triggers>
-                <EventTrigger
-                    RoutedEvent="Grid.MouseEnter">
-                    <BeginStoryboard>
-                        <Storyboard>
-                            <DoubleAnimation
-                                Storyboard.TargetName="titleBar"
-                                Storyboard.TargetProperty="(Grid.Opacity)" 
-                                From="0" To="1" Duration="0:0:0.2"/>
-                        </Storyboard>
-                    </BeginStoryboard>
-                </EventTrigger>
-                <EventTrigger
-                    RoutedEvent="Grid.MouseLeave">
-                    <BeginStoryboard>
-                        <Storyboard>
-                            <DoubleAnimation
-                                Storyboard.TargetName="titleBar"
-                                Storyboard.TargetProperty="(Grid.Opacity)" 
-                                From="1" To="0" Duration="0:0:0.2"/>
-                        </Storyboard>
-                    </BeginStoryboard>
-                </EventTrigger>
-            </Grid.Triggers>
-            <Label
-                Style="{StaticResource labelTitleBar}">
-                Remove
-            </Label>
-            <StackPanel
-                Orientation="Horizontal"
-                HorizontalAlignment="Right">
-                <Button
-                    x:Name="minBtn"
-                    Click="minBtn_Click"
-                    Style="{StaticResource btnTitleBar}"
-                    IsTabStop="False">
-                    -
-                </Button>
-                <Button
-                    x:Name="maxBtn"
-                    Click="maxBtn_Click"
-                    Style="{StaticResource btnTitleBar}"
-                    IsTabStop="False">
-                    o
-                </Button>
-                <Button
-                    x:Name="closeBtn"
-                    Click="closeBtn_Click"
-                    Style="{StaticResource btnTitleBar}"
-                    IsTabStop="False">
-                    x
-                </Button>
-            </StackPanel>
-        </Grid>
-        <!--#endregion-->
-        <StackPanel>
-            <Border Style="{StaticResource borderBot}">
-                <TextBox
-                    x:Name="selectionTxtBox"
-                    KeyDown="selectionTxtBox_KeyDown"
-                    TextChanged="onTextChanged"
-                    Style="{StaticResource txtBox}"/>
-            </Border>
-            <TextBlock
-                x:Name="txtBlock"
-                Style="{StaticResource txtBlock}"/>
-        </StackPanel>
-    </DockPanel>
-</Window>
