+++ MyNetia/Styles.xaml
@@ -124,6 +124,7 @@
     <Style x:Key="borderBot" TargetType="Border">
         <Setter Property="BorderBrush" Value="#510A32"/>
         <Setter Property="BorderThickness" Value="0, 0, 0, 1"/>
+        <Setter Property="Padding" Value="0, 0, 0, 10"/>
     </Style>
 
     <Style x:Key="spV" TargetType="StackPanel">
@@ -193,6 +194,7 @@
     <Style x:Key="btnBase" BasedOn="{StaticResource btnEssential}" TargetType="Button">
         <Setter Property="FontSize" Value="20"/>
         <Setter Property="IsTabStop" Value="False"/>
+        <Setter Property="Margin" Value="10"/>
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="{x:Type Button}">
