+++ MyNetia/TODO.txt
@@ -2,3 +2,4 @@
 - Lien hypertext
 - Avoir un point de vue du résultat final (autre fenetre ? Miniature ? Btn pour Show/Hide ?)
 - Ajouter le fait d avoir des titres (différenciation entre simple texte et titre)
+- Ajouter la suppression de chapitre dans le add/update
